functions and nested loops in C

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -5,15 +5,15 @@
  */
 int main(void)
 {
-	int term1 = 1;
-	int term2 = 2;
-	int next;
+	unsigned long int term1 = 1;
+	unsigned long int term2 = 2;
+	unsigned long int next;
 	int i;
 
-	printf("%d, ", term1);
+	printf("%lu, ", term1);
 	for (i = 1; i < 50; i++)
 	{
-		printf("%d, ", term2);
+		printf("%lu, ", term2);
 		next = term1 + term2;
 		term1 = term2;
 		term2 = next;
